@@ -3,7 +3,7 @@ name = "KiramiBot"
 version = "0.3.5"
 description = "🌟 读作 Kirami，写作星见，简明轻快的聊天机器人应用。"
 authors = [{ name = "Akirami", email = "akiramiaya@outlook.com" }]
-requires-python = ">=3.10,<4.0"
+requires-python = ">=3.12,<4.0"
 license = { text = "AGPL-3.0" }
 readme = "README.md"
 keywords = ["kirami", "bot", "chatbot", "onebot"]
@@ -16,35 +16,32 @@ classifiers = [
     "Programming Language :: Python",
     "Programming Language :: Python :: 3",
     "Programming Language :: Python :: 3 :: Only",
-    "Programming Language :: Python :: 3.10",
-    "Programming Language :: Python :: 3.11",
     "Programming Language :: Python :: 3.12",
+    "Programming Language :: Python :: 3.13",
     "Topic :: Software Development",
     "Typing :: Typed",
 ]
 dependencies = [
     "nonebot2[fastapi]==2.1.3",
     "nonebot-adapter-onebot==2.3.1",
-    "typing-extensions>=4.8.0",
     "APScheduler>=3.10.4",
     "mango-odm>=0.3.2",
     "arrow>=1.3.0",
-    "bidict>=0.22.1",
-    "cashews>=6.3.0",
+    "bidict>=0.23.1",
+    "cashews>=7.3.2",
     "flowery>=0.0.1",
-    "rich>=13.6.0",
+    "rich>=13.9.3",
     "loguru>=0.7.2",
-    "tomli>=2.0.1; python_version < \"3.11\"",
-    "pyyaml>=6.0.1",
-    "playwright>=1.38.0",
-    "httpx[http2]>=0.25.0",
-    "Jinja2>=3.1.2",
-    "tenacity>=8.2.3",
+    "pyyaml>=6.0.2",
+    "playwright>=1.48.0",
+    "httpx[http2]>=0.27.2",
+    "Jinja2>=3.1.4",
+    "tenacity>=9.0.0",
     "filetype>=1.2.0",
     "markdown-it-py[linkify,plugins]>=3.0.0",
     "mdit-py-emoji>=0.1.1",
-    "pygments>=2.16.1",
-    "yarl>=1.9.2",
+    "pygments>=2.18.0",
+    "yarl>=1.16.0",
 ]
 
 [project.urls]
@@ -54,8 +51,8 @@ documentation = "https://kirami.bot/docs"
 
 [tool.pdm.dev-dependencies]
 dev = [
-    "ruff>=0.3.4",
-    "pre-commit>=3.7.0",
+    "ruff>=0.7.0",
+    "pre-commit>=4.0.1"
 ]
 
 [tool.pdm.scripts]
@@ -65,7 +62,7 @@ format = "ruff format"
 post_install = "pre-commit install"
 
 [tool.pyright]
-pythonVersion = "3.10"
+pythonVersion = "3.12"
 pythonPlatform = "All"
 typeCheckingMode = "basic"
 
