@@ -1,8 +1,6 @@
-from typing import Any, TypeVar
+from typing import Any
 
 from mango import Document, Field
-
-T = TypeVar("T")
 
 
 class Group(Document):
@@ -13,7 +11,7 @@ class Group(Document):
     data: dict[str, Any] = Field(default_factory=dict, init=False)
     """群组数据"""
 
-    def get_data(self, name: str, default: T = None) -> T:
+    def get_data[T](self, name: str, default: T = None) -> T:
         return self.data.get(name, default)
 
     async def set_data(self, name: str, value: Any = None) -> None:
