@@ -1,6 +1,6 @@
 import re
 from datetime import datetime, timedelta, tzinfo
-from typing import Any, Literal, NoReturn, TypeVar, overload
+from typing import Any, Literal, NoReturn, overload
 
 from nonebot.dependencies import Dependent
 from nonebot.matcher import Matcher as BaseMatcher
@@ -23,8 +23,6 @@ from kirami.typing import (
 )
 
 # ruff: noqa: PYI021
-
-_T = TypeVar("_T")
 
 class Matcher(BaseMatcher):
     state: State
@@ -53,7 +51,7 @@ class Matcher(BaseMatcher):
     @overload
     def get_argot(self, key: None = None) -> dict[str, Any]: ...
     @overload
-    def get_argot(self, key: str, default: _T) -> Any | _T: ...
+    def get_argot[T](self, key: str, default: T) -> Any | T: ...
     @overload
     def get_argot(self, key: str, default: None = None) -> Any | None: ...
 
