@@ -104,9 +104,10 @@ async def install_browser() -> None:
     from playwright._impl._driver import compute_driver_executable, get_driver_env
 
     logger.info(f"正在安装 {bot_config.browser} 浏览器")
-    driver_executable = compute_driver_executable()
+    driver_executable, driver_cli = compute_driver_executable()
     process = await asyncio.create_subprocess_exec(
         driver_executable,
+        driver_cli,
         "install",
         "--with-deps",
         bot_config.browser,
